add embed stuff to share page too
